+++ packages/fdr-sdk/src/api-definition/index.ts
@@ -9,3 +9,4 @@ export * from "./snippets/curl";
 export * from "./transformer";
 export * from "./types";
 export * from "./unwrap";
+export * from "./url";
