+++ packages/fdr-sdk/src/api-definition/url.ts
@@ -1,8 +1,7 @@
 import { EndpointDefinition, PathPart } from "@fern-api/fdr-sdk/api-definition";
 import { unknownToString } from "@fern-ui/core-utils";
-import { PlaygroundRequestFormState } from "../types";
 
-export function buildQueryParams(queryParameters: Record<string, unknown> | undefined): string {
+function buildQueryParams(queryParameters: Record<string, unknown> | undefined): string {
     if (queryParameters == null) {
         return "";
     }
@@ -15,7 +14,7 @@ export function buildQueryParams(queryParameters: Record<string, unknown> | unde
     return queryParams.size > 0 ? "?" + queryParams.toString() : "";
 }
 
-function buildPath(path: PathPart[], pathParameters?: Record<string, unknown>): string {
+function buildPath(path: PathPart[] = [], pathParameters?: Record<string, unknown>): string {
     return path
         .map((part) => {
             if (part.type === "pathParameter") {
@@ -28,28 +27,42 @@ function buildPath(path: PathPart[], pathParameters?: Record<string, unknown>): 
         .join("");
 }
 
-export function buildRequestUrl(
-    baseUrl: string = "",
-    path: PathPart[] = [],
-    pathParameters: Record<string, unknown> = {},
-    queryParameters: Record<string, unknown> = {},
-): string {
+interface BuildRequestUrlOptions {
+    path?: PathPart[];
+    pathParameters?: Record<string, unknown>;
+    queryParameters?: Record<string, unknown>;
+    baseUrl?: string;
+}
+export function buildRequestUrl({
+    baseUrl = "",
+    path,
+    pathParameters,
+    queryParameters,
+}: BuildRequestUrlOptions): string {
     return baseUrl + buildPath(path, pathParameters) + buildQueryParams(queryParameters);
 }
 
-export function buildEndpointUrl(
-    endpoint: EndpointDefinition | undefined,
-    formState: PlaygroundRequestFormState | undefined,
-    playgroundEnvironment: string | undefined,
-): string {
-    return buildRequestUrl(
-        playgroundEnvironment ??
+interface BuildEndpointUrlOptions {
+    endpoint?: EndpointDefinition;
+    pathParameters?: Record<string, unknown>;
+    queryParameters?: Record<string, unknown>;
+    baseUrl?: string;
+}
+export function buildEndpointUrl({
+    endpoint,
+    pathParameters,
+    queryParameters,
+    baseUrl,
+}: BuildEndpointUrlOptions): string {
+    return buildRequestUrl({
+        baseUrl:
+            baseUrl ??
             (
                 endpoint?.environments?.find((env) => env.id === endpoint.defaultEnvironment) ??
                 endpoint?.environments?.[0]
             )?.baseUrl,
-        endpoint?.path,
-        formState?.pathParameters,
-        formState?.queryParameters,
-    );
+        path: endpoint?.path,
+        pathParameters,
+        queryParameters,
+    });
 }
