+++ packages/ui/app/src/playground/utils/oauth.ts
@@ -1,19 +1,18 @@
 import { APIV1Read } from "@fern-api/fdr-sdk";
-import { EndpointDefinition } from "@fern-api/fdr-sdk/api-definition";
+import { EndpointDefinition, buildEndpointUrl } from "@fern-api/fdr-sdk/api-definition";
 import { unknownToString, visitDiscriminatedUnion } from "@fern-ui/core-utils";
 import jsonpath from "jsonpath";
 import { mapValues } from "lodash-es";
 import { executeProxyRest } from "../fetch-utils/executeProxyRest";
 import { PlaygroundEndpointRequestFormState, ProxyRequest } from "../types";
 import { serializeFormStateBody } from "./serialize";
-import { buildEndpointUrl } from "./url";
 
 export interface OAuthClientCredentialReferencedEndpointLoginFlowProps {
     formState: PlaygroundEndpointRequestFormState;
     endpoint: EndpointDefinition;
     proxyEnvironment: string;
     oAuthClientCredentialsReferencedEndpoint: APIV1Read.OAuthClientCredentialsReferencedEndpoint;
-    playgroundEnvironment: string | undefined;
+    baseUrl: string | undefined;
     setValue: (value: (prev: any) => any) => void;
     closeContainer?: () => void;
     setDisplayFailedLogin?: (value: boolean) => void;
@@ -24,7 +23,7 @@ export const oAuthClientCredentialReferencedEndpointLoginFlow = async ({
     endpoint,
     proxyEnvironment,
     oAuthClientCredentialsReferencedEndpoint,
-    playgroundEnvironment,
+    baseUrl,
     setValue,
     closeContainer,
     setDisplayFailedLogin,
@@ -42,7 +41,12 @@ export const oAuthClientCredentialReferencedEndpointLoginFlow = async ({
     }
 
     const req: ProxyRequest = {
-        url: buildEndpointUrl(endpoint, formState, playgroundEnvironment),
+        url: buildEndpointUrl({
+            endpoint,
+            pathParameters: formState.pathParameters,
+            queryParameters: formState.queryParameters,
+            baseUrl,
+        }),
         method: endpoint.method,
         headers,
         body: await serializeFormStateBody("", endpoint.request?.body, formState.body, false),
